requirements.txt
fastapi
aiohttp
gitpython
aiofiles